# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Laylaty Platform API
  version: 0.1.0-mvp
  description: >
    REST API for the Laylaty MVP (see docs/database/schema.sql and
    docs/backlog/sprint-backlog.md). Every resource here is backed 1:1 by a
    table in the schema. `organization_id` / `store_id` are never accepted
    as client input — they are always derived server-side from the
    authenticated principal's tenant context (blueprint §6).
servers:
  - url: https://api.laylaty.com/v1
    description: Production
  - url: https://api.staging.laylaty.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /merchant/stores/{storeId}/modules:
    parameters:
      - $ref: '#/components/parameters/StoreId'
    get:
      tags: [Stores]
      summary: List enabled Business Template modules for a store (blueprint §9)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StoreModule' }
    patch:
      tags: [Stores]
      summary: Enable/disable modules for a store
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/StoreModule' }
      responses:
        '200': { description: Updated }

  # ---------------------------------------------------------------- Catalog
...
  /stores/{storeId}/reviews:
    parameters:
      - $ref: '#/components/parameters/StoreId'
    get:
      tags: [Reviews]
      summary: List reviews for a store
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Review' }
    post:
      tags: [Reviews]
      summary: Leave a review (must have a completed order at this store)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }

  # ------------------------------------------------------------------ Admin
...
  /admin/audit-logs:
    get:
      tags: [Admin]
      summary: Query the platform audit trail (blueprint §25)
      parameters:
        - name: entity_type
          in: query
          schema: { type: string }
        - name: entity_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditLog' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    StoreId:
      name: storeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    StoreModule:
      type: object
      properties:
        module_key: { type: string, example: booking }
        enabled: { type: boolean }
        config: { type: object }
...
    ReviewInput:
      type: object
      required: [rating]
      properties:
        item_type: { type: string, enum: [product, service, booking, package] }
        item_id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
    Review:
      allOf:
        - $ref: '#/components/schemas/ReviewInput'
        - type: object
          properties:
            id: { type: string, format: uuid }
            customer_id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
    AuditLog:
      type: object
      properties:
        id: { type: string, format: uuid }
        actor_id: { type: string, format: uuid }
        action: { type: string }
        entity_type: { type: string }
        entity_id: { type: string, format: uuid }
        before_state: { type: object }
        after_state: { type: object }
        created_at: { type: string, format: date-time }